requirements.txt
Django==4.2
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
django-cors-headers==4.0.0
djoser==2.2.0
django-templated-mail==1.1.1
requests
